// backend/src/types/controlIva.ts
export interface PurchaseIva {
  cod_emp: number;
  llave: string;
  fecha: string;
  id_tipo_documento: string;
  documento: string;
  cod_proveedor: string;
  periodo_ano: number;
  periodo_mes: number;
  exentas_locales: number;
  exentas_importaciones: number;
  exentas_internaciones: number;
  gravadas_locales: number;
  gravadas_importaciones: number;
  gravadas_internaciones: number;
  no_sujetas: number;
  credito_fiscal: number;
  anticipo_a_cuenta: number;
  iva_retenido: number;
  retencion_a_terceros: number;
  compras_a_excluidos: number;
  rebajas_y_devoluciones: number;
  iva_rebajas_y_devoluciones: number;
  corr_maquina_registradora?: string;
  iva_percibido: number;
  cod_sucursal?: string;
  cod_punto_venta?: string;
  num_control?: string;
  sello_recepcion?: string;
  nom_proveedor?: string;
  registro_proveedor?: string;
  nit_proveedor?: string;
  nom_tipo_documento?: string;
}
// ...
export interface PaginatedResult<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}